+++ packages/search/projects/search-lib/src/lib/search/search.component.ts
@@ -52,8 +52,7 @@ const ALL_LABEL = 'All';
   ],
 })
 export class SearchComponent<T extends IReturnType>
-  implements OnInit, OnDestroy, ControlValueAccessor
-{
+  implements OnInit, OnDestroy, ControlValueAccessor {
   searchBoxInput = '';
   suggestionsDisplay = false;
   categoryDisplay = false;
@@ -93,7 +92,8 @@ export class SearchComponent<T extends IReturnType>
   // emitted when user clicks one of the suggested results (including recent search sugestions)
   @Output() clicked = new EventEmitter<ItemClickedEvent<T>>();
   @Output() searched = new EventEmitter<RecentSearchEvent>();
-  /* emitted when user makes search request (including recent search requests & requests made on change in category from dropdown)
+  /* emitted when user makes search request (including recent search requests
+     & requests made on change in category from dropdown)
   In case of recent search Array of recent Search request result is emitted */
 
   onChange!: (value: string | undefined) => void;
@@ -165,7 +165,8 @@ export class SearchComponent<T extends IReturnType>
       }
     }
     /* need to put default value here and not in contructor
-    because sonar was giving code smell with definite assertion as all these parameters are optional */
+    because sonar was giving code smell with definite assertion as all these
+     parameters are optional */
     const requestParameters: ISearchQuery = {
       match: eventValue.input,
       sources: this._categoryToSourceName(this.category),
@@ -210,9 +211,12 @@ export class SearchComponent<T extends IReturnType>
     }
   }
 
-  // event can be KeyBoardEvent or Event of type 'change' fired on change in value of drop down for category
+  //event can be KeyBoardEvent or Event of type 'change'
+  // fired on change in value of drop down for category
+
   hitSearchApi(event?: Event) {
-    // this will happen only in case user searches something and then erases it, we need to update recent search
+    // this will happen only in case user searches something and
+    // then erases it, we need to update recent search
     if (!this.searchBoxInput) {
       this.suggestions = [];
       this.getRecentSearches();
@@ -240,14 +244,17 @@ export class SearchComponent<T extends IReturnType>
   }
 
   populateValue(suggestion: T, event: MouseEvent) {
-    const value = suggestion[
+    const value = (suggestion[
       this.config.displayPropertyName
-    ] as unknown as string; // converted to string to assign value to searchBoxInput
+    ] as unknown) as string;
+    // converted to string to assign value to searchBoxInput
     this.searchBoxInput = value;
     this.suggestionsDisplay = false;
-    // ngModelChange doesn't detect change in value when populated from outside, hence calling manually
+    // ngModelChange doesn't detect change in value
+    // when populated from outside, hence calling manually
     this.onChange(this.searchBoxInput);
-    // need to do this to show more search options for selected suggestion - just in case user reopens search input
+    // need to do this to show more search options for selected
+    //suggestion - just in case user reopens search input
     this.getSuggestions({input: this.searchBoxInput, event});
     this.clicked.emit({item: suggestion, event});
   }
@@ -258,16 +265,17 @@ export class SearchComponent<T extends IReturnType>
     this.searchBoxInput = value;
     this.suggestionsDisplay = false;
     this.onChange(this.searchBoxInput);
-    // need to do this to show more search options for selected suggestion - just in case user reopens search input
+    // need to do this to show more search options for selected
+    // suggestion - just in case user reopens search input
     this.getSuggestions({input: this.searchBoxInput, event});
     this.focusInput();
     this.showSuggestions();
   }
 
   fetchModelImageUrlFromSuggestion(suggestion: T) {
-    const modelName = suggestion[
-      'source' as unknown as keyof T
-    ] as unknown as string;
+    const modelName = (suggestion[
+      ('source' as unknown) as keyof T
+    ] as unknown) as string;
     let url: string | undefined;
     this.config.models.forEach(model => {
       if (model.name === modelName && model.imageUrl) {
@@ -279,7 +287,7 @@ export class SearchComponent<T extends IReturnType>
 
   boldString(str: T[keyof T] | string, substr: string) {
     const strRegExp = new RegExp(`(${substr})`, 'gi');
-    const stringToMakeBold: string = str as unknown as string;
+    const stringToMakeBold: string = (str as unknown) as string;
     return stringToMakeBold.replace(strRegExp, `<b>$1</b>`);
   }
 
@@ -322,7 +330,8 @@ export class SearchComponent<T extends IReturnType>
     this.suggestions = [];
     this.suggestionsDisplay = true;
     this.focusInput();
-    // ngModelChange doesn't detect change in value when populated from outside, hence calling manually
+    // ngModelChange doesn't detect change in value
+    // when populated from outside, hence calling manually
     this.onChange(this.searchBoxInput);
     this.getRecentSearches();
   }