+++ sandbox/oauth-example/auth-service/src/datasources/index.ts
@@ -1,0 +1,6 @@
+// Copyright (c) 2022 Sourcefuse Technologies
+//
+// This software is released under the MIT License.
+// https://opensource.org/licenses/MIT
+export * from './db.datasource';
+export * from './redis.datasource';