+++ services/user-tenant-service/src/__tests__/acceptance/user-tenant.controller.acceptance.ts
@@ -111,6 +111,7 @@ describe('UserTenant Controller', function () {
       new Tenant({
         name: tenantName,
         key: key,
+        status: 1,
       }),
     );
     const userTenant = await userTenantRepo.create(